requirements.txt
pygit2
python-dotenv
pytz
requests